pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "pypixoo"
version = "2.3.0"
description = "BDD-first Python library for Divoom Pixoo 64"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0",
    "requests~=2.32.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]
[project.urls]
Homepage = "https://github.com/endymion/PyPixoo"
Repository = "https://github.com/endymion/PyPixoo"

[project.scripts]
pypixoo = "pypixoo.cli:main"

[project.optional-dependencies]
dev = [
    "behave~=1.2.6",
    "coverage[toml]>=7.0",
    "playwright>=1.40",
    "pytest>=7.0",
    "sphinx>=7.0",
    "furo>=2024.0.0",
]
browser = [
    "playwright>=1.40",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
source = ["src"]
command_line = "-m behave"

[tool.coverage.report]
exclude_lines = ["pragma: no cover"]

[tool.semantic_release]
branch = "main"
version_toml = ["pyproject.toml:project.version"]
build_command = "pip install build && python -m build"
